Added __all__ to core files module

+++ pygame_menu/widgets/selection/left_arrow.py
@@ -76,9 +76,9 @@ class LeftArrowSelection(ArrowSelection):
         #     margin
         rect = widget.get_rect()
         a = (rect.topleft[0] - self._arrow_size[0] - self._arrow_right_margin,
-             rect.midleft[1] - self._arrow_size[1] / 2 + self._arrow_vertical_offset)
+             int(rect.midleft[1] - self._arrow_size[1] / 2 + self._arrow_vertical_offset))
         b = (rect.midleft[0] - self._arrow_right_margin,
              rect.midleft[1] + self._arrow_vertical_offset)
         c = (rect.bottomleft[0] - self._arrow_size[0] - self._arrow_right_margin,
-             rect.midleft[1] + self._arrow_size[1] / 2 + self._arrow_vertical_offset)
+             int(rect.midleft[1] + self._arrow_size[1] / 2 + self._arrow_vertical_offset))
         super(LeftArrowSelection, self)._draw_arrow(surface, widget, a, b, c)
